refactor: split soil and fullscreen chat helpers
